Fix broken build
Files in the qgym folder were moved and merged with the files in
python/qgym. Furthermore, setup.py was modernized by using pathlib, so
the code becomes more readible, whilts being just as stable.

In the python/qgym/__init__.py explicit exports of the subpackes were
done. This should not be necesary and we should look to remove this in
the future.